service layer tests done

+++ VendingMachine/VendingMachine/Interfaces/IVendingMachineItem.cs
@@ -4,7 +4,7 @@ namespace VendingMachine.Interfaces
     public interface IVendingMachineItem
     {
         public int Quantity { get; }
-        public double Price { get; }
+        public decimal Price { get; }
         public string Name { get;  }
         public string Category { get; }
     }
